Split web app into API and UI blueprints with SSE and Vue UI
