+++ web/templates/index.html
@@ -1,40 +1,86 @@
 <!doctype html>
 <html>
   <head>
-    <title>Mandeye</title>
-    <script>
-      function updateStatus(){
-        fetch('/status')
-          .then(r => r.json())
-          .then(data => {
-            document.getElementById('livox').textContent = JSON.stringify(data.livox, null, 2);
-            document.getElementById('gnss').textContent = JSON.stringify(data.gnss, null, 2);
-            document.getElementById('fs').textContent = JSON.stringify(data.fs || {}, null, 2);
-            document.getElementById('status').textContent = JSON.stringify(data, null, 2);
-          });
-      }
-      function call(url){ fetch(url); }
-      setInterval(updateStatus, 1000);
-    </script>
+    <meta charset="utf-8" />
+    <title>Mandeye Control</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" />
+    <script src="https://cdn.jsdelivr.net/npm/vue@3.4.21/dist/vue.global.prod.js"></script>
   </head>
-  <body onload="updateStatus()">
-    <h1>Mandeye Status</h1>
+  <body>
+    <div id="app" class="container py-4">
+      <h1 class="mb-4">Mandeye Status</h1>
 
-    <h2>Lidar</h2>
-    <pre id="livox"></pre>
+      <div v-if="error" class="alert alert-danger" role="alert">{{ error }}</div>
+      <div v-if="message" class="alert alert-success" role="alert">{{ message }}</div>
 
-    <h2>GNSS</h2>
-    <pre id="gnss"></pre>
+      <div class="mb-3">
+        <button class="btn btn-primary me-2" @click="sendCommand('/api/start_scan')">Start Scan</button>
+        <button class="btn btn-warning me-2" @click="sendCommand('/api/stop_scan')">Stop Scan</button>
+        <button class="btn btn-danger" @click="sendCommand('/api/stopscan')">Trigger Stop Scan</button>
+      </div>
 
-    <h2>Filesystem</h2>
-    <pre id="fs"></pre>
+      <div class="row">
+        <div class="col-md-4">
+          <h2>Lidar</h2>
+          <pre>{{ format(status.livox) }}</pre>
+        </div>
+        <div class="col-md-4">
+          <h2>GNSS</h2>
+          <pre>{{ format(status.gnss) }}</pre>
+        </div>
+        <div class="col-md-4">
+          <h2>Filesystem</h2>
+          <pre>{{ format(status.fs) }}</pre>
+        </div>
+      </div>
 
-    <h2>Full Status</h2>
-    <pre id="status"></pre>
+      <div class="mt-4">
+        <h2>Scan Info</h2>
+        <p>Mode: {{ stream.mode || status.state }}</p>
+        <p>Elapsed: {{ Math.round((stream.dur||0)/1e9) }} s</p>
+        <p>Last Saved File: {{ status.lastLazStatus && status.lastLazStatus.filename || 'N/A' }}</p>
+        <p class="text-danger" v-if="status.state && status.state.includes('ERROR')">Error: {{ status.state }}</p>
+      </div>
+    </div>
 
-    <button onclick="call('/start_scan')">Start Scan</button>
-    <button onclick="call('/stop_scan')">Stop Scan</button>
-    <button onclick="call('/stopscan')">Trigger Stop Scan</button>
-    <button onclick="updateStatus()">Refresh Status</button>
+    <script>
+      const { createApp } = Vue;
+      createApp({
+        data() {
+          return { status: {}, stream: {}, error: '', message: '' };
+        },
+        methods: {
+          format(obj) {
+            return JSON.stringify(obj || {}, null, 2);
+          },
+          updateStatus() {
+            fetch('/api/status')
+              .then(r => r.json())
+              .then(data => { this.status = data; });
+          },
+          sendCommand(url) {
+            fetch(url, { method: 'POST' })
+              .then(r => {
+                if (!r.ok) throw new Error('Command failed');
+                this.message = 'Command successful';
+                setTimeout(() => this.message = '', 3000);
+              })
+              .catch(err => {
+                this.error = err.message;
+                setTimeout(() => this.error = '', 3000);
+              });
+          }
+        },
+        mounted() {
+          this.updateStatus();
+          setInterval(this.updateStatus, 5000);
+
+          const source = new EventSource('/api/stream');
+          source.onmessage = e => {
+            try { this.stream = JSON.parse(e.data); } catch (err) {}
+          };
+        }
+      }).mount('#app');
+    </script>
   </body>
 </html>
