+++ web/ui.py
@@ -1,0 +1,8 @@
+from flask import Blueprint, render_template
+
+ui_bp = Blueprint('ui', __name__)
+
+
+@ui_bp.route('/')
+def index():
+    return render_template('index.html')
